refactor: pull out invitationQueries into separate file
